refactor(settings): 拆出同步域状态 hook、恢复弹窗与同步页签容器
- 新增 settings/sync/：useSyncMeta（同步元信息/模式/墓碑清理）、useSyncProviders（四提供方表单与测试/保存）、useSyncRestore（云端恢复与合并同步）、useRecoveryPassword（恢复密码）、RestoreDialogs（三个恢复弹窗）、SyncTabPane（同步域常驻容器）
- SettingsModal.tsx 2486 -> 1755 行
